archives/skills.old/loreal-api-search/scripts/pyproject.toml
[project]
name = "btdp-api-search-scripts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "google-cloud-secret-manager>=2.20.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
